client/app/requirements.txt
fastapi>=0.95.0
weaviate-client>=3.15.0
sentence-transformers>=2.2.2
uvicorn>=0.21.0fastapi>=0.95.0
weaviate-client>=3.15.0
sentence-transformers>=2.2.2
uvicorn>=0.21.0